chore: fix some bugs
